Movida función middleware de jwt
